br1italia-master/requirements.txt
Django==1.8
MySQL-python==1.2.3
Pillow==2.3.0
argparse==1.2.1
django-appconf==1.0.1
django-image-cropping==0.9.3
django-tinymce==1.5.3
dnspython==1.11.1
easy-thumbnails==2.2
html5lib==0.999
pyOpenSSL==0.13
pycrypto==2.6.1
pyserial==2.6
requests==2.2.1
six==1.5.2
urllib3==1.7.1
virtualenv==12.0.7
wsgiref==0.1.2
zope.interface==4.0.5

